Revolutionize Todo Management with Context API
This commit brings a groundbreaking improvement to our TodoList application by leveraging the power of React's Context API. The CategoryList component has been completely refactored to utilize the TodoListContext provided by TodoListProvider. We have eliminated unnecessary nested TodoListProvider and useTodoList hook, achieving a more elegant and scalable architecture.
