+++ my-app/src/components/Category.tsx
@@ -1,7 +1,5 @@
 import React, {useState, useRef} from "react";
-
-//importing models
-import { Todo } from "../models/todo";
+import { useTodoList } from "../context/TodoListContext";
 
 //importing components
 import TodoList from "./TodoList";
@@ -9,37 +7,27 @@ import NewTodo from "./NewTodo";
 
 
 interface CategoryProps {
-    addTodo: (category: string, text: string) => void;
-    items: Todo[];
-    onStatusUpdate: (category: string, todoId: string, status: 'Completed' | 'Pending') => void;
-    onDelete: (category: string, todoId: string) => void;
     categoryName: string;
-    onTitleEdit: (category: string, newCategoryName: string) => void;
 }
 
-const Category: React.FC<CategoryProps> = (props) => {
+const Category: React.FC<CategoryProps> = ({categoryName}) => {
     const [isTitleEdting, setIsTitleEditing] = useState(false);
-    const categoryTitleInput = useRef<HTMLInputElement>(null)
-
-    const editCategoryNameHandler = (category: string) => {
-        const newCategory = categoryTitleInput.current!.value;
-        props.onTitleEdit(category, newCategory);
-        setIsTitleEditing(prev => false);
-    }
+    const categoryTitleInput = useRef<HTMLInputElement>(null);
+    const { changeCategoryName } = useTodoList();
 
     return (
         <div className="category-container outline outline-2 rounded-lg shadow-xl">
             {isTitleEdting ?
             <>
-            <input ref={categoryTitleInput} type="text" defaultValue={props.categoryName}/>
-            <button onClick={editCategoryNameHandler.bind(null, props.categoryName)}>save</button>
+            <input ref={categoryTitleInput} type="text" defaultValue={categoryName}/>
+            <button onClick={()=> changeCategoryName(categoryName, categoryTitleInput.current!.value)}>save</button>
             </> 
             : 
-            <h3 onClick={()=> {setIsTitleEditing(prev => true)}} className="category-title font-bold mt-2">{props.categoryName}</h3>
+            <h3 onClick={()=> {setIsTitleEditing(prev => true)}} className="category-title font-bold mt-2">{categoryName}</h3>
             }
-            <NewTodo addTodo={props.addTodo.bind(null, props.categoryName)}/>
+            <NewTodo categoryName={categoryName}/>
             <div>
-                <TodoList items={props.items} onStatusUpdate={props.onStatusUpdate.bind(null, props.categoryName)} onDelteTodo={props.onDelete.bind(null, props.categoryName)}/>
+                <TodoList categoryName={categoryName}/>
             </div>
         </div>
     );
